Include documents and metadata with query result (#15)
1. Support non-numerical IDs.
This adds support for user-defined string IDs. They are mapped to individual integer vector IDs that get passed to HNSW.

2. Return document and metadata with query
Text document and metadata can also be retrieved with any query.
Instead of returning a tuple with results and distances, an array of dictionaries are returned which contains id and distance keys by default (and document and metadata if they are selected).